some more r508 fixes
git-svn-id: http://svn.code.sf.net/p/holdingnuts/code/trunk@511 00c8e4f2-bc05-11dd-9892-7df5a783e2f7

+++ src/client/CreateGameDialog.cpp
@@ -162,7 +162,7 @@ QString CreateGameDialog::getName()
 	return editName->text();
 }
 
-float CreateGameDialog::getStake()
+double CreateGameDialog::getStake()
 {
 	return spinStake->value();
 }
@@ -177,12 +177,12 @@ unsigned int CreateGameDialog::getTimeout()
 	return spinTimeout->value();
 }
 
-float CreateGameDialog::getBlindsStart()
+double CreateGameDialog::getBlindsStart()
 {
 	return spinBlindsStart->value();
 }
 
-float CreateGameDialog::getBlindsFactor()
+double CreateGameDialog::getBlindsFactor()
 {
 	return spinBlindsFactor->value();
 }
